mailvine/requirements.txt
appdirs==1.4.4
autopep8==1.5.6
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
distlib==0.3.1
dnspython==2.1.0
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
Pillow==8.2.0
pycodestyle==2.7.0
pycparser==2.20
six==1.15.0
SQLAlchemy==1.4.7
toml==0.10.2
Werkzeug==1.0.1
WTForms==2.3.3
